requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb
openai
python-dotenv
tiktoken